+++ docs/changelog.md
@@ -13,6 +13,8 @@ Here you can find all the changelogs for `hikari-arc`.
 
 - Add `Client.create_task` to make it easier to create "fire and forget" tasks.
 - Add `Client.is_started` and `Client.wait_until_started` for more convenient lifecycle management.
+- Add the ability to pass an already existing injector instance to `Client` via the `injector=` kwarg. If not passed, a new injector will be created by default, like before.
+- Set the client as a type dependency upon instantiation.
 - Make usage of `__slots__` consistent across the library.
 
 ## v1.0.0